pyproject.toml
[project]
# Unfortunately PyPI doesn't allow us to use "last" as the package name.
name = "last-asr"
description = "The LAttice-based Speech Transducer (LAST) library"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [{name = "LAST authors"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]
keywords = ["speech", "jax"]

# pip dependencies of the project
dependencies = [
  "einops>=0.5.0",
  "flax>=0.6.1",
  "jax>=0.3.21",
]

# This is set automatically by flit using `last.__version__`
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/google-research/last"
repository = "https://github.com/google-research/last"
# Other: `documentation`, `changelog`

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install .[dev]`
dev = [
  "absl-py",
  "numpy",
  "optax",
  "pytest",
  "pytest-xdist",
  "pylint>=2.6.0",
  "yapf",
]

[build-system]
requires = ["flit_core >=3.5,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "last"
